+++ src/components/Employees/Employees.tsx
@@ -1,13 +1,13 @@
-import { Stack } from '@mui/material';
+import { Box } from '@mui/material';
 
 import { AddEmployeeButton } from './AddEmployeeButton';
 import { EMTable } from './emTable';
 
 export function Employees() {
   return (
-    <Stack direction="column" spacing="10px">
+    <Box>
       <AddEmployeeButton />
       <EMTable />
-    </Stack>
+    </Box>
   );
 }