+++ src/components/index.ts
@@ -6,3 +6,4 @@ export * from '@app/components/Analytics';
 export * from '@app/components/Employees';
 export * from '@app/components/TransportManagement';
 export * from '@app/components/Auth';
+export * from '@app/components/Organizations';