+++ src/stores/index.ts
@@ -1,14 +1,19 @@
 import { AppStore } from './app';
 import { UsersStore } from './users';
+import { OrganizatonsStore } from './organizations';
+import { TransportStore } from './transport';
 
 export type RootStore = {
   appStore: AppStore;
-  usersStore: UsersStore;
+  orgsStore: OrganizatonsStore;
+  transportStore: TransportStore;
 };
 
 const rootStore: RootStore = {
   appStore: new AppStore(),
   usersStore: new UsersStore(),
+  orgsStore: new OrganizatonsStore(),
+  transportStore: new TransportStore(),
 };
 
 export default rootStore;