+++ CHANGELOG.md
@@ -2,4 +2,59 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.1.0](https://github.com/routelink/client/compare/v1.0.0...v1.1.0) (2024-04-23)
+
+### Features
+
+- add `mobx` to `Organization` component ([#101](https://github.com/routelink/client/issues/101)) ([413b12c](https://github.com/routelink/client/commit/413b12cdf0623a1c2841b8eb9ca57e04532ff06b))
+- add adaptability in `Profile` component ([#89](https://github.com/routelink/client/issues/89)) ([8e2bb26](https://github.com/routelink/client/commit/8e2bb26cf77682cf5aa54a173a78833665c6fee5))
+- add CODEOWNERS ([d4041ef](https://github.com/routelink/client/commit/d4041efcfc5a10cb20741c72c0868105f037be66))
+- add new ci files ([#90](https://github.com/routelink/client/issues/90)) ([955ac1a](https://github.com/routelink/client/commit/955ac1a640a7fe50166d93b2db8f1da318201f91))
+- add theme of app ([42e2a04](https://github.com/routelink/client/commit/42e2a0435837c6e77bd480f6c06e624a46a51ddf))
+- change animation of dialog boxes in `Profile` component ([#91](https://github.com/routelink/client/issues/91)) ([e2224ac](https://github.com/routelink/client/commit/e2224ac64e20964b7aa27f4009d8ab29003c252e))
+- improve styles of dialog boxes in `Profile` component ([#92](https://github.com/routelink/client/issues/92)) ([8eb1879](https://github.com/routelink/client/commit/8eb18791e4a6ac1dfdc722417fc718db68c14307))
+
+### Bug Fixes
+
+- add container padding to `Main` ([#81](https://github.com/routelink/client/issues/81)) ([032ef0e](https://github.com/routelink/client/commit/032ef0e4a89be801055c5582e374594f9a7a55e2))
+- change name column ([#79](https://github.com/routelink/client/issues/79)) ([b149b78](https://github.com/routelink/client/commit/b149b787a3fa868930bd3c535784adedd7f12e17))
+- loading animation ([#102](https://github.com/routelink/client/issues/102)) ([940437d](https://github.com/routelink/client/commit/940437d04adb12b5165fcd2e85bc5fc2eeb57afd))
+- vehicle reg number adding ([#84](https://github.com/routelink/client/issues/84)) ([14755d1](https://github.com/routelink/client/commit/14755d1045e87e40b9fd387191324fe01adc6637))
+
+## [1.0.0](https://github.com/routelink/client/compare/v0.0.1...v1.0.0) (2024-04-17)
+
+### Features
+
+- add `maps` component ([#38](https://github.com/routelink/client/issues/38)) ([126852f](https://github.com/routelink/client/commit/126852fb55dc34a611197bebb0c105bfef055777))
+- add analytics component ([#46](https://github.com/routelink/client/issues/46)) ([61cef0c](https://github.com/routelink/client/commit/61cef0cd018b62ff2bd799f4db71ea8bc8151ac7))
+- add api service ([69c1db6](https://github.com/routelink/client/commit/69c1db6098dfecd150ab7274d989a222f58d67fa))
+- add api service ([#10](https://github.com/routelink/client/issues/10)) ([e73d936](https://github.com/routelink/client/commit/e73d936dd11100cd507341e17ec8f24db94c147c))
+- add auth layout ([#49](https://github.com/routelink/client/issues/49)) ([08382ab](https://github.com/routelink/client/commit/08382ab945c44a070b9e431f884a8d2b7b15bb02))
+- add base layout ([#11](https://github.com/routelink/client/issues/11)) ([cf1b9fb](https://github.com/routelink/client/commit/cf1b9fbceafd9fcabd50ab7f996193d304d62484))
+- add base models ([#32](https://github.com/routelink/client/issues/32)) ([63a735d](https://github.com/routelink/client/commit/63a735d362802d4443ba52cbdcd5b33e1c7f429a))
+- add employee management component ([#48](https://github.com/routelink/client/issues/48)) ([517c38b](https://github.com/routelink/client/commit/517c38b4f32d786bf0eb7ce52af765f31f128d01))
+- add orgs components ([#71](https://github.com/routelink/client/issues/71)) ([ad333cf](https://github.com/routelink/client/commit/ad333cf04d595226f0f279019d9e19c5a20af980))
+- add profile component ([#41](https://github.com/routelink/client/issues/41)) ([668ecdf](https://github.com/routelink/client/commit/668ecdfb82c79df85b18fb1febf81b96b265a83b))
+- add profile component ([#45](https://github.com/routelink/client/issues/45)) ([559d82a](https://github.com/routelink/client/commit/559d82a3cc8e6d8b95876bf97f999fa6c71bb097))
+- add run testing after PR ([#35](https://github.com/routelink/client/issues/35)) ([a3276e5](https://github.com/routelink/client/commit/a3276e58f395e1a1ba794725f463bc90737f5da7))
+- add transport management component ([#40](https://github.com/routelink/client/issues/40)) ([3fe54c5](https://github.com/routelink/client/commit/3fe54c5b1c105a7810c06492a5c5e3b01e65dac8))
+- add users components ([#36](https://github.com/routelink/client/issues/36)) ([cd21de6](https://github.com/routelink/client/commit/cd21de6c82b70069c8538ac1fdb1f41bfa22a29c))
+- improve `Auth` component ([#63](https://github.com/routelink/client/issues/63)) ([35e6f06](https://github.com/routelink/client/commit/35e6f066147538c330c50cacb03e4d8e40899ceb))
+- improve `Menu` component in `Maps` ([#66](https://github.com/routelink/client/issues/66)) ([c989a17](https://github.com/routelink/client/commit/c989a1786017817a8a1cb0c2d0106a94594eb835))
+- increase `z-index` value of layout ([284bf1d](https://github.com/routelink/client/commit/284bf1ddf8abbc06843da49ef5ac1fae138ad9d8))
+- init modal from sidebar ([#26](https://github.com/routelink/client/issues/26)) ([e177316](https://github.com/routelink/client/commit/e177316208e9ad39d1c7aa52c3db140361a62a64))
+- install packages for maps ([#37](https://github.com/routelink/client/issues/37)) ([af3d5b8](https://github.com/routelink/client/commit/af3d5b85353ae15b6e79e7018015c8359b410c19))
+- test ([530bbbc](https://github.com/routelink/client/commit/530bbbc561f4d4154b84bcce93990e4959bc8c3c))
+- update users component ([#61](https://github.com/routelink/client/issues/61)) ([2b19601](https://github.com/routelink/client/commit/2b196013de04c4ee4f5c06440e14ad443f38133c))
+- update users component ([#47](https://github.com/routelink/client/issues/47)) ([f0e4daf](https://github.com/routelink/client/commit/f0e4daf85595f63e08d42e13acd04fe4255576d1))
+- update users components ([#44](https://github.com/routelink/client/issues/44)) ([3af9f91](https://github.com/routelink/client/commit/3af9f916304356d992aac4ae36b4480d3ded9491))
+
+### Bug Fixes
+
+- change addButton style ([#58](https://github.com/routelink/client/issues/58)) ([0ff4c51](https://github.com/routelink/client/commit/0ff4c51e318805bdd1c15b85787014d2ff9ce958))
+- change imports ([6e12830](https://github.com/routelink/client/commit/6e1283057f9af36df1a03a818d516cccae0ff632))
+- delete old profile directory ([2878440](https://github.com/routelink/client/commit/28784409c3d1ee115b6f5bbc83ae70a95cf8a055))
+- hot rewrite unusable component ([8a09eda](https://github.com/routelink/client/commit/8a09eda4dc7989c0759dd883626d71c8cc8ece54))
+- rename `IOrganisation` to `IOrganization` ([61d9ddd](https://github.com/routelink/client/commit/61d9dddf6da43c507560af026ac3c703f1d2c75a))
+
 ### 0.0.1 (2024-03-26)